+++ assets/javascripts/kitten/components/layout/dashboard-layout/stories.js
@@ -50,6 +50,21 @@ const CardHolder = styled.div`
   }
 `
 
+const StyledAvatarWithTextAndBadge = styled(AvatarWithTextAndBadge)`
+  /* FIX AvatarWithTextAndBadge */
+
+  .text--withEllipsis {
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    overflow: hidden;
+  }
+
+  .k-ButtonImageWithText__text {
+    padding-top: 0;
+    padding-bottom: 0;
+  }
+`
+
 const HelpBox = styled.div`
   height: ${pxToRem(80)};
   background-color: rgb(255, 255, 255, 0.05);
@@ -70,7 +85,7 @@ export default {
 
 export const Default = () => {
   const selectedView = select(
-    'content type',
+    'content_type',
     ['flow', 'dashboard', 'table'],
     'flow',
   )
@@ -86,9 +101,10 @@ export const Default = () => {
         closeLabel: 'Fermer le menu',
       }}
       quickAccessLinkText="Accéder au contenu"
+      fullHeightContent={selectedView === 'flow'}
     >
       <DashboardLayout.Header>
-        <AvatarWithTextAndBadge>
+        <StyledAvatarWithTextAndBadge>
           <AvatarWithTextAndBadge.Image
             src="/kitten.jpg"
             alt=""
@@ -111,60 +127,64 @@ export const Default = () => {
               Prêt a être partagé avec mon coach
             </Text>
           </AvatarWithTextAndBadge.Text>
-        </AvatarWithTextAndBadge>
+        </StyledAvatarWithTextAndBadge>
       </DashboardLayout.Header>
       <DashboardLayout.SideContent>
         {({ closeSideBar }) => {
           return (
             <DashboardMenu>
               <DashboardMenu.Item
-                href="#"
+                href="?id=layout-dashboardlayout--default&knob-content_type=dashboard&viewMode=story"
+                isActive={
+                  window.location.search ===
+                  '?id=layout-dashboardlayout--default&knob-content_type=dashboard&viewMode=story'
+                }
                 icon={() => <HomeIcon color="currentColor" />}
-                onClick={e => {
-                  e.preventDefault()
-                  closeSideBar()
-                }}
+                onClick={() => closeSideBar()}
               >
                 Accueil
               </DashboardMenu.Item>
               <DashboardMenu.Item
-                href="#"
-                isActive
+                href="?id=layout-dashboardlayout--default&knob-content_type=table&viewMode=story"
+                isActive={
+                  window.location.search ===
+                  '?id=layout-dashboardlayout--default&knob-content_type=table&viewMode=story'
+                }
                 icon={() => <TagIcon color="currentColor" />}
-                onClick={e => {
-                  e.preventDefault()
-                  closeSideBar()
-                }}
+                onClick={() => closeSideBar()}
               >
                 Contreparties
               </DashboardMenu.Item>
               <DashboardMenu.Item
-                href="#"
+                href="?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=a"
+                isActive={
+                  window.location.search ===
+                  '?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=a'
+                }
                 icon={() => <FilterIcon color="currentColor" />}
-                onClick={e => {
-                  e.preventDefault()
-                  closeSideBar()
-                }}
+                onClick={() => closeSideBar()}
               >
                 Paramètres
               </DashboardMenu.Item>
               <DashboardMenu.Item
-                href="https://www.kisskissbankbank.com"
+                href="?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story"
+                isActive={
+                  window.location.search ===
+                  '?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story'
+                }
                 icon={() => <FileIcon color="currentColor" />}
-                onClick={e => {
-                  e.preventDefault()
-                  closeSideBar()
-                }}
+                onClick={() => closeSideBar()}
               >
                 Page Projet
               </DashboardMenu.Item>
               <DashboardMenu.Item
-                href="https://www.kisskissbankbank.com"
+                href="?id=layout-dashboardlayout--default&knob-content_type=flow&knob-loading=true&viewMode=story"
+                isActive={
+                  window.location.search ===
+                  '?id=layout-dashboardlayout--default&knob-content_type=flow&knob-loading=true&viewMode=story'
+                }
                 icon={() => <LoudspeakerIcon color="currentColor" />}
-                onClick={e => {
-                  e.preventDefault()
-                  closeSideBar()
-                }}
+                onClick={() => closeSideBar()}
               >
                 Marketing
               </DashboardMenu.Item>
@@ -173,30 +193,32 @@ export const Default = () => {
                 icon={() => <ShieldIcon color="currentColor" />}
               >
                 <DashboardMenu.Item
-                  href="#"
-                  onClick={e => {
-                    e.preventDefault()
-                    closeSideBar()
-                  }}
+                  href="?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=1"
+                  isActive={
+                    window.location.search ===
+                    '?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=1'
+                  }
+                  onClick={() => closeSideBar()}
                 >
                   Destinataire des fonds
                 </DashboardMenu.Item>
                 <DashboardMenu.Item
-                  href="#"
-                  onClick={e => {
-                    e.preventDefault()
-                    closeSideBar()
-                  }}
-                  isActive
+                  href="?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=2"
+                  isActive={
+                    window.location.search ===
+                    '?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=2'
+                  }
+                  onClick={() => closeSideBar()}
                 >
                   Confirmation d'identité
                 </DashboardMenu.Item>
                 <DashboardMenu.Item
-                  href="#"
-                  onClick={e => {
-                    e.preventDefault()
-                    closeSideBar()
-                  }}
+                  href="?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=3"
+                  isActive={
+                    window.location.search ===
+                    '?id=layout-dashboardlayout--default&knob-content_type=flow&viewMode=story&custom=3'
+                  }
+                  onClick={() => closeSideBar()}
                 >
                   Documents justificatifs
                 </DashboardMenu.Item>
@@ -207,7 +229,7 @@ export const Default = () => {
       </DashboardLayout.SideContent>
       <DashboardLayout.SideFooter>
         <HelpBox>
-          <AvatarWithTextAndBadge>
+          <StyledAvatarWithTextAndBadge>
             <AvatarWithTextAndBadge.Image src="/kitten.jpg" alt="" />
 
             <AvatarWithTextAndBadge.Text>
@@ -229,7 +251,7 @@ export const Default = () => {
                 Contacter votre coach
               </Text>
             </AvatarWithTextAndBadge.Text>
-          </AvatarWithTextAndBadge>
+          </StyledAvatarWithTextAndBadge>
         </HelpBox>
       </DashboardLayout.SideFooter>